openshift.ks: only backup fw conf if it exists
configure_firewall: Only try to read and backup
/etc/sysconfig/system-config-firewall if it exists.

This commit fixes bug 1169997.

+++ enterprise/install-scripts/amazon/openshift-amz.sh
@@ -2937,8 +2937,7 @@ configure_firewall()
 {
   # Disable lokkit.
   conf='/etc/sysconfig/system-config-firewall'
-  [[ "$(< "$conf")" != --disabled ]] &&
-    mv "$conf" "${conf}.bak"
+  [[ -e "$conf" && "$(< "$conf")" != --disabled ]] && mv "$conf" "${conf}.bak"
   echo '--disabled' > "$conf"
 
   # Configure iptables.
